# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Career Scout API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /online-discovery/run:
    post:
      operationId: runOnlineDiscovery
      tags: [onlineDiscovery]
      summary: Find new online job opportunities matched to the current profile
      responses:
        "200":
          description: Discovery run completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OnlineDiscoveryRunResult"
        "400":
          description: Profile needs experience or skills before discovery can run
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /online-discovery/settings:
    put:
      operationId: updateOnlineDiscoverySettings
      tags: [onlineDiscovery]
      summary: Update online job discovery settings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OnlineDiscoverySettings"
      responses:
        "200":
          description: Updated discovery status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OnlineDiscoveryStatus"
...
  /imap/connect:
    post:
      operationId: connectImap
      tags: [imap]
      summary: Connect an IMAP account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConnectImapBody"
      responses:
        "200":
          description: IMAP connected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImapConnected"
        "400":
          description: Missing required fields
        "422":
          description: Connection failed
...
  /imap/sync:
    post:
      operationId: syncImap
      tags: [imap]
      summary: Trigger a manual IMAP sync
      responses:
        "200":
          description: Sync result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImapSyncResult"
        "400":
          description: No IMAP connection configured
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    OnlineDiscoverySettings:
      type: object
      properties:
        scheduleHours:
          type: ["number", "null"]
        minimumMatchScore:
          type: integer
          minimum: 1
          maximum: 100
      required:
        - scheduleHours
        - minimumMatchScore
    OnlineDiscoveryCriteria:
      type: object
      properties:
        roleTitles:
          type: array
          items:
            type: string
        skills:
          type: array
          items:
            type: string
        locations:
          type: array
          items:
            type: string
        remotePreferences:
          type: array
          items:
            type: string
      required:
        - roleTitles
        - skills
        - locations
        - remotePreferences
    OnlineDiscoveryStatus:
      type: object
      properties:
        source:
          type: string
        scheduleHours:
          type: ["number", "null"]
        minimumMatchScore:
          type: integer
        lastRunAt:
          type: ["string", "null"]
          format: date-time
        nextRunAt:
          type: ["string", "null"]
          format: date-time
        lastFound:
          type: integer
        lastImported:
          type: integer
        lastDuplicates:
          type: integer
        lastError:
          type: ["string", "null"]
        criteria:
          $ref: "#/components/schemas/OnlineDiscoveryCriteria"
      required:
        - source
        - minimumMatchScore
        - lastFound
        - lastImported
        - lastDuplicates
        - criteria
    OnlineDiscoveryRunResult:
      allOf:
        - $ref: "#/components/schemas/OnlineDiscoveryStatus"
        - type: object
          properties:
            fetched:
              type: integer
            considered:
              type: integer
            imported:
              type: integer
            duplicates:
              type: integer
            matchedExisting:
              type: integer
          required:
            - fetched
            - considered
            - imported
            - duplicates
            - matchedExisting
...
    ConnectImapBody:
      type: object
      properties:
        host:
          type: string
        port:
          type: integer
        username:
          type: string
        password:
          type: string
        tls:
          type: boolean
      required:
        - host
        - port
        - username
        - password
    ImapConnected:
      type: object
      properties:
        connected:
          type: boolean
        host:
          type: string
        port:
          type: integer
        username:
          type: string
        tls:
          type: boolean
      required:
        - connected
        - host
        - port
        - username
        - tls
    ImapSyncResult:
      type: object
      properties:
        synced:
          type: integer
      required:
        - synced